matched Actor00100_Fn01388 9
Same obstacle-bearing walk as Actor00100_Fn00508, with a 16-slot scratch
(Actor00100AvoidScratch16) and a flags field gating blocked for kind 0x10000.
The carve sits above the early-out guard, which is what the target's block
layout shows, so that path leaks the scratch.

+++ include/actors/actor_400100_facing.h
@@ -5,7 +5,7 @@
 #include "actors/actor_400100_update.h"
 #include "main/wipsys.h"
 
-s32                          Actor00100_Fn01388(GsCOORDINATE2*, GpRec18*, s32, SVECTOR*);
+s32                          Actor00100_Fn01388(GsCOORDINATE2*, GpRec18*, s16, SVECTOR*);
 extern char                  Actor00100_D000D4;
 extern Actor00100AnimCommand Actor00100_D1B9AC;
 extern SVECTOR               Actor00100_D1BA90;
